refactor: remove unusing sellerchat (#215)
